fix: close streams, deflake bundled-copy test, longer error snackbar
- FileUtils.copy now wraps streams in .use { } so a mid-copy IOException
  closes both descriptors; without it, every recurring failure leaked two
  FDs (the prior crash-and-die masked it).
- ShareFeatureTest deletes the leftover bundled-share file in setUp so the
  copy-failure test reaches its mock regardless of sibling-test order.
- Error snackbars in LandingScreen now use SnackbarDuration.Long: copy is
  70+ chars in es-AR and includes a CTA, ~4s wasn't enough.

+++ app/src/main/java/com/github/barriosnahuel/vossosunboton/ui/home/LandingScreen.kt
@@ -213,9 +213,11 @@ private suspend fun shareWithFeedback(
 ) {
     val errorRes = ShareFeature.instance.share(context, sound, surface)
     if (errorRes != null) {
+        // Long (~7s) instead of Short (~4s): error copy includes a call-to-action ("Free up some space",
+        // "Try installing one") that needs reading time, and our error strings are 70+ chars in es-AR.
         snackbarHostState.showSnackbar(
             message = context.getString(errorRes),
-            duration = SnackbarDuration.Short,
+            duration = SnackbarDuration.Long,
         )
     }
 }
